Update notebooklm extension
- add new files
- refinement with pr review -2
